+++ src/controls/defaultControls.ts
@@ -1,107 +1,20 @@
-import { changeWidth } from './changeWidth';
-import { Control } from './Control';
-import { rotationStyleHandler, rotationWithSnapping } from './rotate';
-import { scaleCursorStyleHandler, scalingEqually } from './scale';
-import {
-  scaleOrSkewActionName,
-  scaleSkewCursorStyleHandler,
-  scalingXOrSkewingY,
-  scalingYOrSkewingX,
-} from './scaleSkew';
+import { FabricObject } from '../shapes/Object/FabricObject';
+import { Textbox } from '../shapes/Textbox';
+import { defaultControls, textboxDefaultControls } from './commonControls';
 
-// use this function if you want to generate new controls for every instance
-export const createObjectDefaultControls = () => ({
-  ml: new Control({
-    x: -0.5,
-    y: 0,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingXOrSkewingY,
-    getActionName: scaleOrSkewActionName,
-  }),
+FabricObject.prototype.controls = {
+  ...(FabricObject.prototype.controls || {}),
+  ...defaultControls,
+};
 
-  mr: new Control({
-    x: 0.5,
-    y: 0,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingXOrSkewingY,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  mb: new Control({
-    x: 0,
-    y: 0.5,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingYOrSkewingX,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  mt: new Control({
-    x: 0,
-    y: -0.5,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingYOrSkewingX,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  tl: new Control({
-    x: -0.5,
-    y: -0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  tr: new Control({
-    x: 0.5,
-    y: -0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  bl: new Control({
-    x: -0.5,
-    y: 0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  br: new Control({
-    x: 0.5,
-    y: 0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  mtr: new Control({
-    x: 0,
-    y: -0.5,
-    actionHandler: rotationWithSnapping,
-    cursorStyleHandler: rotationStyleHandler,
-    offsetY: -40,
-    withConnection: true,
-    actionName: 'rotate',
-  }),
-});
-
-export const createResizeControls = () => ({
-  mr: new Control({
-    x: 0.5,
-    y: 0,
-    actionHandler: changeWidth,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionName: 'resizing',
-  }),
-  ml: new Control({
-    x: -0.5,
-    y: 0,
-    actionHandler: changeWidth,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionName: 'resizing',
-  }),
-});
-
-export const createTextboxDefaultControls = () => ({
-  ...createObjectDefaultControls(),
-  ...createResizeControls(),
-});
-
-export const defaultControls = createObjectDefaultControls();
+if (Textbox) {
+  // this is breaking the prototype inheritance, no time / ideas to fix it.
+  // is important to document that if you want to have all objects to have a
+  // specific custom control, you have to add it to Object prototype and to Textbox
+  // prototype. The controls are shared as references. So changes to control `tr`
+  // can still apply to all objects if needed.
+  Textbox.prototype.controls = {
+    ...(Textbox.prototype.controls || {}),
+    ...textboxDefaultControls,
+  };
+}
